requirements.txt.txt
streamlit
openai
python-dotenv
pypdf
faiss-cpu
numpy